Fix bug (https://github.com/VorontsovIE/sarus/issues/2) with pvalue >= 1 which doesn't force sarus to print scores for all positions.
The problem was due to floating point arithmetics which allow score for matrix
to have score less than the worst one. It can occur for example for computations
in superalphabet. Right now all P-values greater than 1.0 will be treated
in a special way - so that all positions will be printed with guarantees.